# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "API documentation for managing profiles and admin authentication",
        "contact": {
            "email": "jean@test.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local server"
        }
    ],
    "paths": {
        "/admin/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Register a new admin",
                "description": "Creates a new admin and returns their details",
                "operationId": "registerAdmin",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "jean@test.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password123"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Admin created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "email": {
                                            "type": "string",
                                            "example": "jean@test.com"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/admin/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Admin login",
                "description": "Authenticates an admin and returns a token",
                "operationId": "loginAdmin",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "jean@test.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Token returned",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string",
                                            "example": "Bearer token"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid credentials"
                    }
                }
            }
        },
        "/profiles": {
            "post": {
                "tags": [
                    "Profiles"
                ],
                "summary": "Create a profile",
                "description": "Creates a new profile for a user",
                "operationId": "createProfile",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "firstname": {
                                        "type": "string",
                                        "example": "Jean"
                                    },
                                    "lastname": {
                                        "type": "string",
                                        "example": "Test"
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "active"
                                    },
                                    "image": {
                                        "description": "Profile image",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Profile created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "firstname": {
                                            "type": "string",
                                            "example": "Jean"
                                        },
                                        "lastname": {
                                            "type": "string",
                                            "example": "Test"
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "active"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Invalid data"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/profiles/active": {
            "get": {
                "tags": [
                    "Profiles"
                ],
                "summary": "Get all active profiles",
                "description": "Retrieve a list of all active profiles",
                "operationId": "getActiveProfiles",
                "responses": {
                    "200": {
                        "description": "List of active profiles",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Profile"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/profiles/{id}": {
            "put": {
                "tags": [
                    "Profiles"
                ],
                "summary": "Update a profile",
                "description": "Updates a profile by ID",
                "operationId": "updateProfile",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Profile ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "firstname": {
                                        "type": "string",
                                        "example": "Jean"
                                    },
                                    "lastname": {
                                        "type": "string",
                                        "example": "Test"
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "active"
                                    },
                                    "image": {
                                        "description": "Profile image",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Profile updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "firstname": {
                                            "type": "string",
                                            "example": "Jean"
                                        },
                                        "lastname": {
                                            "type": "string",
                                            "example": "Test"
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "active"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "422": {
                        "description": "Invalid data"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Profiles"
                ],
                "summary": "Delete a profile",
                "description": "Deletes a profile by ID",
                "operationId": "deleteProfile",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Profile ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Profile deleted sucessfully"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Admin": {
                "description": "Admin model",
                "properties": {
                    "email": {
                        "description": "Email of the administrator",
                        "type": "string"
                    },
                    "password": {
                        "description": "Password of the administrator",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "Profile": {
                "description": "Profile model",
                "properties": {
                    "firstname": {
                        "description": "Firstname of the profile",
                        "type": "string"
                    },
                    "lastname": {
                        "description": "Lastname of the profile",
                        "type": "string"
                    },
                    "image": {
                        "description": "Link to the profile picture of the profile",
                        "type": "string"
                    },
                    "status": {
                        "description": "Status of the profile",
                        "type": "string"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "tags": [
        {
            "name": "Authentication",
            "description": "Authentication"
        },
        {
            "name": "Profiles",
            "description": "Profiles"
        }
    ],
    "security": [
        {
            "sanctum": []
        }
    ]
}